refactor(bom): move download bom button to util
